relocate universal components
