+++ apps/app/src/client/components/ReactMarkdownComponents/Header.tsx
@@ -5,12 +5,12 @@ import type EventEmitter from 'events';
 import { useRouter } from 'next/router';
 import type { Element } from 'react-markdown/lib/rehype-filter';
 
+import { NextLink } from '~/components/ReactMarkdownComponents/NextLink';
 import {
   useIsGuestUser, useIsReadOnlyUser, useIsSharedUser, useShareLinkId,
 } from '~/stores-universal/context';
 import loggerFactory from '~/utils/logger';
 
-import { NextLink } from './NextLink';
 
 import styles from './Header.module.scss';
