+++ apps/app/src/models/.eslintrc.js
