+++ apps/app/src/utils/.eslintrc.js
@@ -5,6 +5,7 @@ module.exports = {
     'no-restricted-imports': ['error', {
       patterns: [
         '~/client/',
+        'client/',
       ],
     }],
   },
